feat(rice-warehouse): add sales-to-outbound flow, non-paddy purchasin…

+++ Backend.Application/DTOs/Locations/CreateLocationDto.cs
@@ -9,10 +9,10 @@ public class CreateLocationDto
     public string? ShelfRow { get; set; }
     public string? ShelfLevel { get; set; }
     public string? SlotCode { get; set; }
-    public int? MaxCapacity { get; set; }
+    public decimal? MaxCapacity { get; set; }
     public string? Description { get; set; }
     public bool IsActive { get; set; }
-    public int CurrentOccupancy { get; set; }
+    public decimal CurrentOccupancy { get; set; }
     public int? AllowedCategoryId { get; set; }
     public int Priority { get; set; }
     public bool IsQuarantine { get; set; }
