+++ Backend.Domain/Aggregates/LocationAggregate.cs
@@ -11,8 +11,8 @@ public class LocationAggregate
     public string? ShelfRow { get; set; }
     public string? ShelfLevel { get; set; }
     public string? SlotCode { get; set; }
-    public int? MaxCapacity { get; set; }
-    public int CurrentOccupancy { get; set; }
+    public decimal? MaxCapacity { get; set; }
+    public decimal CurrentOccupancy { get; set; }
     public string? Description { get; set; }
     public bool IsActive { get; set; }
     public DateTime CreatedDate { get; set; }
